Extract types out of D.IDE.Core.Shake
to avoid circular module dependencies
